requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
attrs==24.2.0
ccxt==4.4.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
filelock==3.16.0
frozenlist==1.4.1
fsspec==2024.9.0
idna==3.8
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.1.0
networkx==3.3
numpy==2.1.1
pandas==2.2.2
pycares==4.4.0
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
setuptools==74.1.2
six==1.16.0
sympy==1.13.2
threadpoolctl==3.5.0
torch==2.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
yarl==1.11.1
